requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
dj-database-url
python-dotenv
psycopg[binary]
gunicorn
